backend/requirements.txt
fastapi==0.109.0
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.9.0
groq==0.7.2
pymongo==4.7.0
passlib==1.7.4
python-jose==3.3.0
